# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/crear-admin": {
      "post": {
        "operationId": "AdminController_createAdmin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Un admin crea un nuevo admin",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/usuarios": {
      "get": {
        "operationId": "AdminController_findAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "un admin obtiene todos los usuarios",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/usuarios/{id}/rol": {
      "patch": {
        "operationId": "AdminController_changeUserRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarRol"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Un admin cambia el rol de cualquier usuario",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/tiendas": {
      "get": {
        "operationId": "AdminController_getAllTiendas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Un admin ve todas tiendas",
        "tags": [
          "Admin"
        ]
      }
    },
    "/gemas": {
      "post": {
        "operationId": "GemaController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearGenericoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Un admin crea una gema",
        "tags": [
          "Gemas"
        ]
      },
      "get": {
        "operationId": "GemaController_findAll",
        "parameters": [
          {
            "name": "nombre",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todos las gemas con filtro opcional por nombre",
        "tags": [
          "Gemas"
        ]
      }
    },
    "/gemas/{id}": {
      "get": {
        "operationId": "GemaController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener una gema por ID",
        "tags": [
          "Gemas"
        ]
      },
      "patch": {
        "operationId": "GemaController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarGenericoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar una gema por ID",
        "tags": [
          "Gemas"
        ]
      },
      "delete": {
        "operationId": "GemaController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar una gema por ID",
        "tags": [
          "Gemas"
        ]
      }
    },
    "/materiales": {
      "post": {
        "operationId": "MaterialController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearGenericoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Un admin crea un material",
        "tags": [
          "Materiales"
        ]
      },
      "get": {
        "operationId": "MaterialController_findAll",
        "parameters": [
          {
            "name": "nombre",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todos los materiales con filtro opcional por nombre",
        "tags": [
          "Materiales"
        ]
      }
    },
    "/materiales/{id}": {
      "get": {
        "operationId": "MaterialController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener un material por ID",
        "tags": [
          "Materiales"
        ]
      },
      "patch": {
        "operationId": "MaterialController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarGenericoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar un material por ID",
        "tags": [
          "Materiales"
        ]
      },
      "delete": {
        "operationId": "MaterialController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar un material por ID",
        "tags": [
          "Materiales"
        ]
      }
    },
...
    "/estilos": {
      "post": {
        "operationId": "EstiloController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearGenericoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Un admin crea un estilo",
        "tags": [
          "Estilos"
        ]
      },
      "get": {
        "operationId": "EstiloController_findAll",
        "parameters": [
          {
            "name": "nombre",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todos los estilos con filtro opcional por nombre, se puede consultar asi tambien para todas las caracteristicas /estilos?nombre=mo",
        "tags": [
          "Estilos"
        ]
      }
    },
    "/estilos/{id}": {
      "get": {
        "operationId": "EstiloController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener un estilo por ID",
        "tags": [
          "Estilos"
        ]
      },
      "patch": {
        "operationId": "EstiloController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarGenericoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar un estilo por ID",
        "tags": [
          "Estilos"
        ]
      },
      "delete": {
        "operationId": "EstiloController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar un estilo por ID",
        "tags": [
          "Estilos"
        ]
      }
    },
...
    "/tiendas": {
      "get": {
        "operationId": "TiendaController_getAllPublicas",
        "parameters": [
          {
            "name": "nombre",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Listar todos los estilos con filtro opcional por nombre, se puede consultar asi tambien para todas las caracteristicas /tiendas?nombre=mo",
        "tags": [
          "Tienda"
        ]
      }
    },
...
  },
  "info": {
    "title": "MarketJoyas",
    "description": "Documentacion de endpoints de marketjoyas",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CrearAdminDto": {
        "type": "object",
        "properties": {
          "nombre_completo": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "dni": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "contraseña": {
            "type": "string",
            "minLength": 6
          }
        },
        "required": [
          "nombre_completo",
          "email",
          "dni",
          "telefono",
          "contraseña"
        ]
      },
      "CambiarRol": {
        "type": "object",
        "properties": {
          "rol": {
            "type": "object"
          }
        },
        "required": [
          "rol"
        ]
      },
      "CrearGenericoDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "description": "Nombre de la entidad (unico)",
            "example": "x"
          }
        },
        "required": [
          "nombre"
        ]
      },
      "ActualizarGenericoDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "description": "Nombre de la entidad (unico)",
            "example": "Moderno"
          }
        }
      },
...
    }
  }
}